Refactored RESTSubscriber::enable and ::register_enable_route integration tests.
