+++ tests/Integration/RESTfulTestCase.php
@@ -42,7 +42,30 @@ abstract class RESTfulTestCase extends TestCase {
 			];
 		}
 
-		$request = new WP_Rest_Request( 'PUT', '/wp-rocket/v1/rocketcdn/disable' );
+		return $this->doRestRequest( $body_params, '/wp-rocket/v1/rocketcdn/disable' );
+	}
+
+	/**
+	 * Runs the RESTful endpoint which invokes WordPress to run in an integrated fashion. Callback will be fired.
+	 *
+	 * @param array $body_params Optional. Array of body parameters.
+	 *
+	 * @return array a response packet.
+	 */
+	protected function requestEnableEndpoint( array $body_params = [] ) {
+		if ( empty( $body_params ) ) {
+			$body_params = [
+				'email' => '',
+				'key'   => '',
+				'url'   => 'https://rocketcdn.me',
+			];
+		}
+
+		return $this->doRestRequest( $body_params, '/wp-rocket/v1/rocketcdn/enable' );
+	}
+
+	protected function doRestRequest( array $body_params, $route ) {
+		$request = new WP_Rest_Request( 'PUT', $route );
 		$request->set_header( 'Content-Type', 'application/x-www-form-urlencoded' );
 		$request->set_body_params( $body_params );
